posts features almost complete

+++ admin/src/lib/api.ts
@@ -1,7 +1,18 @@
-import { AnimeType, movieType, SeriesType } from "../common/types";
+import { AnimeType, movieType, PostDtoType, SeriesType } from "../common/types";
+import {
+  imdbValidationSchema,
+  myAnimeListValidationSchema,
+  PostIdValidationSchema,
+} from "./validation.zod";
 
 export async function getMovie(imdb_id: string) {
-  const response = await fetch(`http://localhost:3001/movie/${imdb_id}`);
+  //validating
+  const validated = imdbValidationSchema.safeParse({ imdb_id });
+
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+  const response = await fetch(
+    `http://localhost:3001/movie/${validated.data.imdb_id}`
+  );
 
   const data = await response.json();
 
@@ -13,6 +24,10 @@ export async function getMovie(imdb_id: string) {
 }
 
 export async function findOrAddMovie(imdb_id: string) {
+  //Validating
+  const validated = imdbValidationSchema.safeParse({ imdb_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   const response = await fetch(
     `http://localhost:3001/movie/find-add?imdb_id=${imdb_id}`
   );
@@ -25,9 +40,13 @@ export async function findOrAddMovie(imdb_id: string) {
   return await response.json();
 }
 
-export async function updateMovie(id: string, updateDto: movieType) {
+export async function updateMovie(imdb_id: string, updateDto: movieType) {
+  //Validating
+  const validated = imdbValidationSchema.safeParse({ imdb_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   await new Promise((res) => setTimeout(() => res(""), 3000));
-  const response = await fetch(`http://localhost:3001/movie/${id}`, {
+  const response = await fetch(`http://localhost:3001/movie/${imdb_id}`, {
     method: "PATCH",
     body: JSON.stringify(updateDto),
     headers: {
@@ -54,7 +73,10 @@ export async function getAllMovies() {
 }
 
 export async function deleteMovie(imdb_id: string) {
-  console.log("checking imdb_id", imdb_id);
+  //Validating
+  const validated = imdbValidationSchema.safeParse({ imdb_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   await new Promise((res, rej) => setTimeout(() => res("s"), 3000));
   const response = await fetch(`http://localhost:3001/movie/${imdb_id}`, {
     method: "DELETE",
@@ -69,8 +91,11 @@ export async function deleteMovie(imdb_id: string) {
 }
 
 //series
-
 export async function getSeries(imdb_id: string) {
+  //Validating
+  const validated = imdbValidationSchema.safeParse({ imdb_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   const response = await fetch(`http://localhost:3001/series/${imdb_id}`);
 
   const data = await response.json();
@@ -82,6 +107,10 @@ export async function getSeries(imdb_id: string) {
   return data;
 }
 export async function findOrAddSeries(imdb_id: string) {
+  //Validating
+  const validated = imdbValidationSchema.safeParse({ imdb_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   const response = await fetch(
     `http://localhost:3001/series/find-add?imdb_id=${imdb_id}`
   );
@@ -96,6 +125,10 @@ export async function findOrAddSeries(imdb_id: string) {
 }
 
 export async function updatesSeries(imdb_id: string, updateDto: SeriesType) {
+  //Validating
+  const validated = imdbValidationSchema.safeParse({ imdb_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   await new Promise((res, rej) => setTimeout(() => res("hi"), 3000));
   const response = await fetch(`http://localhost:3001/series/${imdb_id}`, {
     method: "PATCH",
@@ -115,6 +148,10 @@ export async function updatesSeries(imdb_id: string, updateDto: SeriesType) {
 }
 
 export async function deleteSeries(imdb_id: string) {
+  //Validating
+  const validated = imdbValidationSchema.safeParse({ imdb_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   await new Promise((res, rej) => setTimeout(() => res("s"), 3000));
   const response = await fetch(`http://localhost:3001/series/${imdb_id}`, {
     method: "DELETE",
@@ -124,8 +161,6 @@ export async function deleteSeries(imdb_id: string) {
   if (!response.ok) {
     throw new Error(data.message ?? "faild to fetch data");
   }
-
-  console.log("logging delete data", data);
 
   return data;
 }
@@ -146,6 +181,10 @@ export async function getAllSeries() {
 //anime
 
 export async function getAnime(mal_id: string) {
+  //Validating
+  const validated = myAnimeListValidationSchema.safeParse({ mal_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   const response = await fetch(`http://localhost:3001/anime/${mal_id}`);
 
   const data = await response.json();
@@ -157,6 +196,10 @@ export async function getAnime(mal_id: string) {
   return data;
 }
 export async function findOrAddAnime(mal_id: string) {
+  //Validating
+  const validated = myAnimeListValidationSchema.safeParse({ mal_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   const response = await fetch(
     `http://localhost:3001/anime/find-add?mal_id=${mal_id}`
   );
@@ -174,6 +217,10 @@ export async function updateAnime(
   mal_id: string,
   updateDto: Partial<AnimeType>
 ) {
+  //Validating
+  const validated = myAnimeListValidationSchema.safeParse({ mal_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   await new Promise((res, rej) => setTimeout(() => res("hi"), 3000));
   const response = await fetch(`http://localhost:3001/anime/${mal_id}`, {
     method: "PATCH",
@@ -206,6 +253,10 @@ export async function getAllAnimes() {
 }
 
 export async function deleteAnime(mal_id: string) {
+  //Validating
+  const validated = myAnimeListValidationSchema.safeParse({ mal_id });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
   await new Promise((res, rej) => setTimeout(() => res("s"), 3000));
   const response = await fetch(`http://localhost:3001/anime/${mal_id}`, {
     method: "DELETE",
@@ -217,6 +268,56 @@ export async function deleteAnime(mal_id: string) {
   }
 
   console.log("logging delete data", data);
+
+  return data;
+}
+
+//post
+export async function addPost(postDto: PostDtoType) {
+  const response = await fetch(`http://localhost:3001/post`, {
+    method: "Post",
+    body: JSON.stringify(postDto),
+    headers: {
+      "Content-Type": "application/json",
+    },
+  });
+
+  const data = await response.json();
+
+  if (!response.ok) {
+    throw new Error(data.message ?? "faild to fetch data");
+  }
+
+  return data;
+}
+
+export async function getAllPosts() {
+  await new Promise((res) => setTimeout(() => res(""), 3000));
+
+  const response = await fetch(`http://localhost:3001/post`);
+
+  const data = await response.json();
+  if (!response.ok) {
+    throw new Error(data.message ?? "Failed to fetch data");
+  }
+
+  return data;
+}
+
+export async function deletePost(postId: string) {
+  //Validating
+  const validated = PostIdValidationSchema.safeParse({ postId });
+  if (!validated.success) throw new Error(validated.error.issues[0].message);
+
+  await new Promise((res, rej) => setTimeout(() => res("s"), 3000));
+  const response = await fetch(`http://localhost:3001/post/${postId}`, {
+    method: "DELETE",
+  });
+  const data = await response.json();
+
+  if (!response.ok) {
+    throw new Error(data.message ?? "faild to fetch data");
+  }
 
   return data;
 }
